Remove the menu from selectionMatrix experiments
There is still a bit more clutter in the process than is ideal (a decoding
screen and closing screen), but that allows for time to look at the results
that the decoder shows by itself. In the final version, we should probably
transition straight from the Calibrating to the Quit state.